+++ test/unit/pulled_search/get_archive_files.py
@@ -36,6 +36,7 @@ class UnitTest(unittest.TestCase):
 
     Methods:
         setUp
+        test_one_month
         test_two_months
 
     """