+++ CHANGELOG.md
@@ -11,6 +11,7 @@ The format is based on "Keep a Changelog".  This project adheres to Semantic Ver
 - main: Added -F option to opt_val_list variable.
 
 ### Changed
+- get_archive_files: Set end datetime to pulldate if exists otherwise use current datetime.
 - process_docid: Determine if pulldate exist for -a (archive) option and pass to get_archive_files call.
 - cleanup_files: Added destination directory to the log entry when archiving the file.
 - recall_search2: Passed pulldate to the search_docid call.